+++ hwcomposer_window.cpp
@@ -488,7 +488,8 @@ void HWComposerNativeWindow::allocateBuffers()
 
         if (b->status) {
             b->common.decRef(&b->common);
-            fprintf(stderr,"WARNING: %s: allocated only %lu buffers out of %d\n", __PRETTY_FUNCTION__, m_bufList.size(), m_bufferCount);
+            fprintf(stderr,"WARNING: %s: allocated only %d buffers out of %d\n", __PRETTY_FUNCTION__,
+                    (uint32_t) m_bufList.size(), m_bufferCount);
             break;
         }
